feat: Implement comprehensive project improvements
This commit implements a wide range of improvements based on a thorough review of the YINI project, addressing recommendations from the `IMPROVEMENTS.md` and `REVIEW.md` files.

The key enhancements include:

1.  **Documentation:**
    - A new `docs/cookbook.md` file has been created to provide practical, copy-and-paste examples for common C# and C++ use cases.
    - The main documentation index (`docs/README.md`) has been updated to include a link to the new "Cookbook," making it easily discoverable.

2.  **Build System:**
    - The root `CMakeLists.txt` has been cleaned up by removing redundant global C++ standard settings, deferring to the modern per-target configuration.

3.  **Code Quality:**
    - Resolved all compiler warnings to improve code robustness and maintainability:
        - Changed index types from `int` to `size_t` in the `Lexer` and `Parser` to fix sign-compare warnings.
        - Added `[[maybe_unused]]` to intentionally unused parameters and variables in the `Interpreter` and `YiniManager`.
        - Corrected missing field initializers for `Token` structs in the `Interpreter`.
        - Fixed an ignored attribute warning in `CliTest.cpp` by using an explicit function pointer type for the `std::unique_ptr` deleter.

+++ src/Lexer/Lexer.h
@@ -31,8 +31,8 @@ namespace YINI
         std::string m_source;
         std::string m_filepath;
         std::vector<Token> m_tokens;
-        int m_start = 0;
-        int m_current = 0;
+        size_t m_start = 0;
+        size_t m_current = 0;
         int m_line = 1;
         int m_column = 1;
         int m_start_column = 1;
